fixed blocking issue and frame buffer issue with attacks (hopefully) and other issues
did not fix the down/up tilt in air issue (movement) would have to overhaul the animation system to do so. idk if it's my computer but sometimes the attacks act wonky? also an issue with jumping and trying to do a tilt attack too early (works fine for air combos?)